+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Move docs directory to xmake-docs repo
 * Improve install and uninstall actions and support DESTDIR and PREFIX envirnoment variables
 * Optimize to detect flags
+* Add `COLORTERM=nocolor` to disable color output
 
 ### Bugs fixed
 
@@ -360,6 +361,7 @@
 * 移除docs目录，将其放置到独立xmake-docs仓库中，减少xmake.zip的大小，优化下载安装的效率
 * 改进安装和卸载脚本，支持DESTDIR和PREFIX环境变量设置
 * 通过缓存优化flags探测，加速编译效率
+* 添加`COLORTERM=nocolor`环境变量开关，禁用彩色输出
 
 ### Bugs修复